requirements.txt
requests>=2.25.1
pandas>=1.1.5
python-dotenv>=0.21.0